Handling kernel weight copying at framework level
On mobile platform, we utilize layout optimization on kernel weights and data.
Previously, OpenCLDataTransfer only handled data tensor layout optimization on
copying data from host to device Image2D. To identify weights and what layout
it should be in, we need to know the tensor usage on copying. We attach the
usage information to NodeArg and Tensor to achieve it.

Additional consideration would be using transformer to optimize the general
kernels in graph into more specialized ones. Then the layout rearrangement can
be done in the transformation and the copying can be handled with
CopyBuffer2DToImage2D. However, this requires us to come up with many contrib
OPs definitions along with their optimization passes. This can be done latter.